beginning ESM transition: Ceramic libraries, Next.js, & TypeScript configuration 🇭🇰
