+++ package.json
@@ -89,11 +89,11 @@
     "standard-version": "9.3.2",
     "ts-jest": "28.0.7",
     "tsdx": "0.14.1",
-    "typescript": "4.6.2",
+    "typescript": "^4.8.4",
     "wait-on": "6.0.1"
   },
   "resolutions": {
-    "typescript": "4.6.2"
+    "typescript": "4.8.4"
   },
   "browserslist": [
     "defaults",
