+++ packages/backend/package.json
@@ -18,14 +18,14 @@
   "author": "MetaFam",
   "license": "ISC",
   "dependencies": {
-    "@ceramicnetwork/http-client": "1.5.7",
-    "@ceramicnetwork/stream-caip10-link": "1.2.9",
-    "@datamodels/identity-accounts-web": "0.1.2",
-    "@datamodels/identity-profile-basic": "0.1.2",
-    "@glazed/datamodel": "0.2.0",
-    "@glazed/devtools": "0.1.5",
-    "@glazed/did-datastore": "0.2.3",
-    "@glazed/tile-loader": "0.1.3",
+    "@ceramicnetwork/http-client": "^2.5.0",
+    "@ceramicnetwork/stream-caip10-link": "^2.3.0",
+    "@datamodels/identity-accounts-web": "^0.2.0",
+    "@datamodels/identity-profile-basic": "^0.2.0",
+    "@glazed/datamodel": "^0.3.1",
+    "@glazed/devtools": "^0.2.0",
+    "@glazed/did-datastore": "^0.3.2",
+    "@glazed/tile-loader": "^0.2.1",
     "@graphql-tools/schema": "8.3.2",
     "@metafam/discord-bot": "0.1.0",
     "@metafam/utils": "1.0.1",
