+++ packages/utils/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.1",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
+  "type": "module",
   "license": "MIT",
   "private": true,
   "scripts": {
@@ -14,21 +15,21 @@
     "test": "jest --passWithNoTests"
   },
   "dependencies": {
-    "@datamodels/identity-profile-basic": "0.1.2",
     "@ethersproject/address": "5.6.0",
     "bignumber.js": "9.0.2",
     "cids": "1.1.9",
     "ethers": "5.6.9",
     "imgix-core-js": "2.3.2",
     "js-base64": "3.7.2",
-    "key-did-provider-ed25519": "1.1.0",
     "node-fetch": "3.2.1",
     "sourcecred": "0.11.0",
     "uuid": "8.3.2"
   },
   "devDependencies": {
-    "@ceramicnetwork/http-client": "1.5.7",
-    "@datamodels/identity-accounts-crypto": "0.1.2",
-    "key-did-resolver": "2.0.4"
+    "@ceramicnetwork/http-client": "^2.5.0",
+    "@datamodels/identity-accounts-crypto": "^0.2.0",
+    "@datamodels/identity-profile-basic": "^0.2.0",
+    "key-did-provider-ed25519": "^2.0.1",
+    "key-did-resolver": "^2.1.3"
   }
 }
