+++ packages/web/.eslintrc.json
@@ -1,4 +1,5 @@
 {
+  "extends": ["plugin:@next/next/recommended"],
   "rules": {
     "no-console": [
       "error",
