+++ packages/web/next-env.d.ts
@@ -1,6 +1,5 @@
 /// <reference types="next" />
-/// <reference types="next/types/global" />
-declare module '*.png';
-declare module '*.jpg';
-declare module '*.jpeg';
-declare module '*.svg';
+/// <reference types="next/image-types/global" />
+
+// NOTE: This file should not be edited
+// see https://nextjs.org/docs/basic-features/typescript for more information.
