+++ packages/web/package.json
@@ -1,7 +1,8 @@
 {
   "name": "@metafam/web",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "concurrently \"next\" \"yarn generate --watch\"",
     "build": "next build",
@@ -12,19 +13,17 @@
     "prepare": "yarn generate"
   },
   "dependencies": {
-    "@3id/connect": "0.2.5",
-    "@ceramicnetwork/3id-did-resolver": "1.4.16",
-    "@ceramicnetwork/http-client": "1.5.7",
-    "@ceramicnetwork/stream-caip10-link": "1.2.9",
-    "@datamodels/identity-profile-basic": "0.1.2",
+    "@ceramicnetwork/http-client": "^2.5.0",
+    "@ceramicnetwork/stream-caip10-link": "^2.3.0",
+    "@datamodels/identity-profile-basic": "^0.2.0",
     "@ethersproject/address": "5.6.0",
-    "@glazed/devtools": "0.1.5",
-    "@glazed/did-datastore": "0.2.3",
-    "@glazed/tile-loader": "0.1.3",
     "@honeybadger-io/js": "^4.3.1",
     "@honeybadger-io/react": "^4.4.1",
     "@honeybadger-io/webpack": "^4.3.1",
-    "@metafam/ds": "0.1.0",
+    "@glazed/devtools": "^0.2.0",
+    "@glazed/did-datastore": "^0.3.2",
+    "@glazed/tile-loader": "^0.2.1",
+    "@metafam/ds": "0.2.0",
     "@metafam/utils": "1.0.1",
     "@quest-chains/sdk": "^0.1.12",
     "@react-spring/web": "9.4.3",
@@ -54,7 +53,7 @@
     "logrocket": "^3.0.1",
     "markdown-it": "13.0.1",
     "moment": "2.29.4",
-    "next": "10.2.3",
+    "next": "^12.3.1",
     "next-images": "1.8.4",
     "next-transpile-modules": "9.0.0",
     "next-urql": "3.3.2",
@@ -82,6 +81,7 @@
     "web3modal": "1.9.5"
   },
   "devDependencies": {
+    "@next/eslint-plugin-next": "^12.3.1",
     "@types/busboy": "1.5.0",
     "@types/deep-equal": "1.0.1",
     "@types/dompurify": "^2.3.3",
